#50 Updating package name
