--- src/main/java/com/jos/dem/jugoterapia/webflux/config/ApplicationConfig.java
@@ -1,29 +1,0 @@
-/*
-  Copyright 2021 Jose Morales <joseluis.delacruz@gmail.com>
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-package com.jos.dem.jugoterapia.webflux.config;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Getter
-@Setter
-@Component
-@ConfigurationProperties("api")
-public class ApplicationConfig {
-
-  private String host;
-  private String bucketUrl;
-}
