--- src/main/java/com/jos/dem/jugoterapia/webflux/config/Bootstrap.java
@@ -1,39 +1,0 @@
-/*
-  Copyright 2021 Jose Morales <joseluis.delacruz@gmail.com>
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-package com.jos.dem.jugoterapia.webflux.config;
-
-import com.jos.dem.jugoterapia.webflux.repository.CategoryRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.context.event.ApplicationReadyEvent;
-import org.springframework.context.ApplicationListener;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class Bootstrap implements ApplicationListener<ApplicationReadyEvent> {
-
-  private final CategoryRepository categoryRepository;
-
-  @Override
-  public void onApplicationEvent(final ApplicationReadyEvent event) {
-    log.info("Trying to load categories...");
-    validateCategories();
-  }
-
-  private void validateCategories() {
-    categoryRepository.findAll().forEach(category -> log.info("Category: {}", category.getName()));
-  }
-}
