--- src/main/java/com/jos/dem/jugoterapia/webflux/controller/HealthController.java
@@ -1,37 +1,0 @@
-/*
-  Copyright 2021 Jose Morales <joseluis.delacruz@gmail.com>
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-package com.jos.dem.jugoterapia.webflux.controller;
-
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Tag(name = "knows how to respond to health checks")
-@Slf4j
-@RestController
-@RequestMapping("/health")
-public class HealthController {
-
-  @Parameter(name = "ping", description = "Ping message", required = true)
-  @GetMapping(value = "/{ping}", produces = MediaType.APPLICATION_JSON_VALUE)
-  public String check(@PathVariable("ping") String ping) {
-    log.info(ping);
-    return "pong";
-  }
-}
