+++ src/main/java/com/josdem/fruitypedia/service/BeverageService.java
@@ -1,0 +1,27 @@
+/*
+  Copyright 2021 Jose Morales <joseluis.delacruz@gmail.com>
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+  http://www.apache.org/licenses/LICENSE-2.0
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+*/
+
+package com.josdem.fruitypedia.service;
+
+import com.josdem.fruitypedia.model.Beverage;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface BeverageService {
+  Optional<Beverage> findById(Integer beverageId);
+
+  List<Beverage> findByCategoryId(Integer categoryId);
+
+  List<Beverage> findByIngredientKeyword(String keyword);
+}
