--- src/test/java/com/jos/dem/jugoterapia/webflux/CategoryControllerTest.java
@@ -1,74 +1,0 @@
-/*
-  Copyright 2021 Jose Morales <joseluis.delacruz@gmail.com>
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-package com.jos.dem.jugoterapia.webflux;
-
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.Matchers.isA;
-import static org.springframework.http.MediaType.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import java.util.ArrayList;
-import lombok.RequiredArgsConstructor;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.test.web.servlet.MockMvc;
-
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@AutoConfigureMockMvc
-@RequiredArgsConstructor(onConstructor_ = @Autowired)
-class CategoryControllerTest {
-
-  private final MockMvc mockMvc;
-
-  @Test
-  @DisplayName("Should get all categories")
-  void shouldGetCategories() throws Exception {
-    mockMvc
-        .perform(get("/categories/"))
-        .andExpect(status().isOk())
-        .andExpect(content().contentTypeCompatibleWith(APPLICATION_JSON_VALUE))
-        .andExpect(jsonPath("$.*", isA(ArrayList.class)));
-  }
-
-  @Test
-  @DisplayName("Should get categories in spanish")
-  void shouldGetCategoriesByLanguage() throws Exception {
-    mockMvc
-        .perform(get("/categories/{language}", "es"))
-        .andExpect(status().isOk())
-        .andExpect(content().contentTypeCompatibleWith(APPLICATION_JSON_VALUE))
-        .andExpect(jsonPath("$.*", isA(ArrayList.class)))
-        .andExpect(jsonPath("$[0].name", equalTo("Curativos")))
-        .andExpect(jsonPath("$[1].name", equalTo("Energizantes")))
-        .andExpect(jsonPath("$[2].name", equalTo("Saludables")))
-        .andExpect(jsonPath("$[3].name", equalTo("Estimulantes")));
-  }
-
-  @Test
-  @DisplayName("Should get categories by id")
-  public void shouldGetBeveragesByCategory() throws Exception {
-    mockMvc
-        .perform(get("/categories/{id}/beverages", 1))
-        .andExpect(status().isOk())
-        .andExpect(content().contentTypeCompatibleWith(APPLICATION_JSON_VALUE))
-        .andExpect(jsonPath("$.*", isA(ArrayList.class)));
-  }
-}
