--- src/test/java/com/jos/dem/jugoterapia/webflux/HealthControllerTest.java
@@ -1,42 +1,0 @@
-/*
-  Copyright 2021 Jose Morales <joseluis.delacruz@gmail.com>
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-package com.jos.dem.jugoterapia.webflux;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import com.jos.dem.jugoterapia.webflux.controller.HealthController;
-import lombok.RequiredArgsConstructor;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.test.web.servlet.MockMvc;
-
-@WebMvcTest(HealthController.class)
-@RequiredArgsConstructor(onConstructor_ = @Autowired)
-class HealthControllerTest {
-
-  private final MockMvc mockMvc;
-
-  @Test
-  @DisplayName("Should get pong")
-  void shouldGetPong() throws Exception {
-    mockMvc
-        .perform(get("/health/{ping}", "ping"))
-        .andExpect(status().isOk())
-        .andExpect(content().string("pong"));
-  }
-}
