--- src/test/java/com/jos/dem/jugoterapia/webflux/LanguageResolverTest.java
@@ -1,53 +1,0 @@
-/*
-  Copyright 2021 Jose Morales <joseluis.delacruz@gmail.com>
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  http://www.apache.org/licenses/LICENSE-2.0
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-package com.jos.dem.jugoterapia.webflux;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import com.jos.dem.jugoterapia.webflux.util.LanguageResolver;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-class LanguageResolverTest {
-
-  private final LanguageResolver languageResolver = new LanguageResolver();
-
-  @Test
-  @DisplayName("Should select English if not language")
-  void shouldSelectEnglishIfNoLanguage() {
-    String language = languageResolver.resolve(null);
-    assertEquals("en", language);
-  }
-
-  @Test
-  @DisplayName("Should select English if empty language")
-  void shouldSelectEnglishIfEmptyLanguage() {
-    String language = languageResolver.resolve("");
-    assertEquals("en", language);
-  }
-
-  @Test
-  @DisplayName("Should select English if other language")
-  void shouldSelectEnglishIfOtherLanguage() {
-    String language = languageResolver.resolve("ru");
-    assertEquals("en", language);
-  }
-
-  @Test
-  @DisplayName("Should select Spanish language")
-  void shouldSelectSpanishLanguage() {
-    String language = languageResolver.resolve("es");
-    assertEquals("es", language);
-  }
-}
